1433: move user class to userdata
