+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/cards/CanonicalJson.kt
@@ -1,6 +1,6 @@
 package app.ehrenamtskarte.backend.cards
 
-import app.ehrenamtskarte.backend.user.KoblenzUser
+import app.ehrenamtskarte.backend.userdata.KoblenzUser
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.protobuf.Descriptors
