+++ backend/src/test/kotlin/app/ehrenamtskarte/backend/cards/Argon2IdHasherTest.kt
@@ -2,7 +2,7 @@ package app.ehrenamtskarte.backend.verification
 import Argon2IdHasher
 import app.ehrenamtskarte.backend.common.utils.Environment
 import app.ehrenamtskarte.backend.common.webservice.KOBLENZ_PEPPER_SYS_ENV
-import app.ehrenamtskarte.backend.user.KoblenzUser
+import app.ehrenamtskarte.backend.userdata.KoblenzUser
 import io.mockk.every
 import io.mockk.mockkObject
 import kotlin.test.Test
