+++ backend/src/test/kotlin/app/ehrenamtskarte/backend/helper/ExampleUser.kt
@@ -1,5 +1,5 @@
 package app.ehrenamtskarte.backend.helper
 
-import app.ehrenamtskarte.backend.user.KoblenzUser
+import app.ehrenamtskarte.backend.userdata.KoblenzUser
 
 val koblenzTestUser = KoblenzUser("Karla Koblenz", 12213, "123K")
